feat: Introduce password visibility toggle, localize login page to Uzbek, and enhance Hemis login logging.

+++ backend/app/modules/quiz_process/schemas.py
@@ -32,4 +32,4 @@ class EndQuizResponse(BaseModel):
     total_questions: int
     correct_answers: int
     wrong_answers: int
-    grade: float
+    grade: int
